refactor(condition): Moved persistense directory into infrastructure
