+++ module/condition/src/Infrastructure/Persistence/Projector/DbalConditionSetDeletedEventProjector.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Condition\Persistence\Dbal\Projector;
+namespace Ergonode\Condition\Infrastructure\Persistence\Projector;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
@@ -15,7 +15,7 @@ use Ergonode\Condition\Domain\Event\ConditionSetDeletedEvent;
 
 /**
  */
-class ConditionSetDeletedEventProjector
+class DbalConditionSetDeletedEventProjector
 {
     private const TABLE = 'condition_set';
