+++ module/condition/src/Infrastructure/Persistence/Query/DbalConditionSetQuery.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Condition\Persistence\Dbal\Query;
+namespace Ergonode\Condition\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
